Add betty 100-prime_factor.c

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,19 +7,15 @@
  */
 int main(void)
 {
-
-long unsigned int n = 612852475143, a = (int) sqrt(n);
-
+unsigned short int n = 612852475143, a = (int)sqrt(n);
 while (1)
 {
-
 if (n % a == 0)
 {
 printf("%lu \n", n / a);
 break;
 }
 a--;
-
 }
 
 return (0);
